backend/requirements.txt
# Enhanced LLM Agent Shopping Assistant v2.1.0 Requirements
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
openai==1.3.0
zhipuai==2.1.5.20250825
sqlalchemy==2.0.23
# sqlite3 is part of Python standard library, no need to install
pillow==10.1.0
speechrecognition==3.10.0
pyttsx3==2.90
websockets==12.0
aiofiles==23.2.1
httpx==0.25.2

# 记忆系统和向量数据库
faiss-cpu==1.7.4
numpy==1.24.3
sentence-transformers==2.2.2
scikit-learn==1.3.0

# RAG系统
langchain==0.1.0
langchain-community==0.0.10
langchain-core==0.1.5
langchain-openai==0.0.2
unstructured==0.12.0
pypdf==3.17.0
docx2txt==0.8
tiktoken==0.5.2

# 多Agent系统
asyncio-mqtt==0.16.1
celery==5.3.4
redis==5.0.1

# 文档处理
PyPDF2==3.0.1
python-docx==1.1.0
markdown==3.5.1
beautifulsoup4==4.12.2
lxml==4.9.3

# 数据处理
pandas==2.1.4
numpy==1.24.3

# 网购助手相关
requests==2.31.0
aiohttp==3.9.1
beautifulsoup4==4.12.2
selenium==4.15.0
opencv-python==4.8.0
# hashlib is part of Python standard library
# PIL is provided by Pillow (already installed above)

# 网络爬虫系统
scrapy==2.11.0
playwright==1.40.0
scrapy-playwright==0.0.36
scrapy-rotating-proxies==0.6.2
fake-useragent==1.4.0
lxml==4.9.3
html2text==2020.1.16
itemadapter==0.8.0
Twisted==23.10.0